Remove componentWillReceiveProps from LocalGraph
